chore: restore v0.8.1 source
