+++ .gitignore
@@ -1,5 +1,4 @@
 node_modules/
+dist/
 .DS_Store
 npm-debug.log*
-/grid-calendar-card.js
-/grid-calendar-card.js.map
