+++ README.md
@@ -1,10 +1,10 @@
-# Grid Calendar Card
+# Multi-Calendar Grid Card
 
 A modern 7-day, time-grid calendar card for Home Assistant. Overlay multiple calendar entities, stack overlapping events, keep all-day events tidy in a pill row, and (optionally) show **daily weather in each day header**.
 
 **Tech:** Lit + TypeScript • Works in normal dashboards and kiosk displays.
 
-> Footer shows the running version. Example: `Grid Calendar Card v0.8.1`.
+> Footer shows the running version. Example: `Multi-Calendar Grid Card v0.8.1`.
 
 ---
 
@@ -18,37 +18,36 @@ A modern 7-day, time-grid calendar card for Home Assistant. Overlay multiple cal
 - **Internationalization support** (new in v0.8.1)
 - Remembered scroll offset (optional)
 - Lightweight, no external dependencies
+
 Tested on **Home Assistant 2025.8**. Should work on 2024.12+.
+
 ---
+
 ## Installation
+
 ### 1) Download the release asset
-Grab `grid-calendar-card.js` from the latest GitHub Release and place it at:
+Grab `multi-calendar-grid-card.js` from the latest GitHub Release and place it at:
 
 ```
-/config/www/grid-calendar-card/grid-calendar-card.js
+/config/www/multi-calendar-grid-card/multi-calendar-grid-card.js
 ```
 
 ### 2) Add a Lovelace resource
 **Settings → Dashboards → (⋮) Resources → + Add resource**
 
-- URL: `/local/grid-calendar-card/grid-calendar-card.js`
+- URL: `/local/multi-calendar-grid-card/multi-calendar-grid-card.js`
 - Resource type: **JavaScript Module**
 
 > Tip: After updates, hard-refresh the browser (Ctrl/Cmd-Shift-R) to bust cache.
 
 ---
 
-## Configuration
-
-### Configuration (UI)
-The card includes a basic editor and appears in Home Assistant's card picker. Use **Add Card → Grid Calendar Card** and configure options through the form. Advanced setups can still be done in YAML as documented below.
-
-### Configuration (YAML)
+## Configuration (YAML)
 
 Add the card to a view:
 
 ```yaml
-type: custom:grid-calendar-card
+type: custom:multi-calendar-grid-card
 entities:
   - entity: calendar.nuesken_family_gmail
     name: Family
@@ -59,21 +58,27 @@ entities:
   - entity: calendar.auri_nuesken_gmail_com
     name: Auri
     color: '#03a9f4'
-```
+
 # Time-grid & layout
-view_start_time: '07:00:00'
-view_end_time: '22:00:00'
-view_slot_minutes: 60
+first_day: 1               # 0=Sun … 6=Sat (used if start_today: false)
+start_today: true          # NEW: start the 7-day view at “today”
+slot_min_time: '07:00:00'
+slot_max_time: '22:00:00'
+slot_minutes: 60
 locale: de               # optional language override
 time_format: 24          # optional time format (12 or 24)
 px_per_min: 0.8
 height_vh: 80
+header_compact: false
 show_now_indicator: true
+show_all_day: true
 remember_offset: true
 data_refresh_minutes: 5
 
 # Weather in day headers (optional)
 weather_entity: weather.integra_langsbau_1_3
+weather_days: 7            # default 7
+weather_compact: false     # false = show icon + hi/low; true = tighter
 ```
 
 ---
@@ -83,17 +88,23 @@ weather_entity: weather.integra_langsbau_1_3
 | Key                    | Type            | Default | Description |
 |------------------------|-----------------|---------|-------------|
 | `entities`             | list            | —       | Calendars to overlay. Each item: `{ entity, name?, color? }`. |
-| `view_start_time`      | `HH:MM:SS`      | `07:00:00` | Earliest visible hour. |
-| `view_end_time`        | `HH:MM:SS`      | `22:00:00` | Latest visible hour. |
-| `view_slot_minutes`    | number          | `30`    | Minor grid step in minutes (1–180). |
-| `locale`               | `'en'`/`'de'`   | HA      | Language override (defaults to Home Assistant). |
-| `time_format`          | `'12'`/`'24'`   | `24`    | Hour format override. |
+| `first_day`            | number (0–6)    | `1`     | Week start (Mon=1). Used only when `start_today: false`. |
+| `start_today`          | boolean         | `true`  | **NEW**. When `true`, the 7-day view starts at today (rolling window). |
+| `slot_min_time`        | `HH:MM:SS`      | `07:00:00` | Earliest visible hour. |
+| `slot_max_time`        | `HH:MM:SS`      | `22:00:00` | Latest visible hour. |
+| `slot_minutes`         | number          | `30`    | Minor grid step in minutes (1–180). |
+| `locale`               | string          | HA      | Language override (defaults to Home Assistant). |
+| `time_format`          | `'12'`/`'24'`   | HA      | Hour format override. |
 | `px_per_min`           | number          | `1.6`   | Vertical scale: pixels per minute. |
 | `height_vh`            | number          | `80`    | Scroll area height in viewport units. |
+| `header_compact`       | boolean         | `false` | Smaller top header. |
 | `show_now_indicator`   | boolean         | `true`  | Red “now” line when viewing the current week. |
+| `show_all_day`         | boolean         | `true`  | Show all-day pill row. |
 | `remember_offset`      | boolean         | `true`  | Persist vertical scroll between reloads. |
 | `data_refresh_minutes` | number          | `5`     | Re-fetch calendars every N minutes (1–60). |
 | `weather_entity`       | string          | —       | Optional `weather.*` entity to show daily weather in headers. |
+| `weather_days`         | number          | `7`     | Days of weather to show (capped by your forecast). |
+| `weather_compact`      | boolean         | `false` | Compact weather display in headers. |
 
 ---
 
@@ -120,7 +131,7 @@ npm run build   # produces bundle(s) in dist/
 ```
 
 For release, we typically upload a stable-named bundle:
-- `dist/grid-calendar-card.js` (+ optional `.map`)
+- `dist/multi-calendar-grid-card.js` (+ optional `.map`)
 
 ---
 
@@ -138,7 +149,6 @@ The most useful docs live in the `/docs` directory:
 - **[0001: Forecast data source](docs/adr/0001-forecast-data-source.md)** – Why we use the service/REST/attributes fallback order.
 - **[0002: Week start = today](docs/adr/0002-week-start-today.md)** – Rationale for starting the 7-day grid from the current day.
 - **[0003: Weather rendering in headers](docs/adr/0003-weather-rendering-in-headers.md)** – Design choices for minimal, consistent weather display.
-- **[0005: Remove legacy options](docs/adr/0005-remove-legacy-options.md)** – Cleanup and renaming of configuration.
 
 See **[CONTRIBUTING.md](CONTRIBUTING.md)** for guidelines on filing issues and sending PRs.
