--- docs/images
@@ -1,1 +1,0 @@
-
