+++ hacs.json
@@ -1,6 +1,6 @@
 {
-  "name": "Grid Calendar Card",
-  "filename": "dist/grid-calendar-card.js",
+  "name": "Multi-Calendar Grid Card",
+  "filename": "multi-calendar-grid-card.js",
   "render_readme": true
 }
