+++ src/locales/de.ts
@@ -10,21 +10,4 @@ export default {
   aria_next_week: "Nächste Woche",
   aria_today: "Zur aktuellen Woche",
   failed_load_prefix: "Fehler beim Laden:",
-  editor_calendars: "Kalender",
-  editor_add_calendar: "Kalender hinzufügen",
-  editor_entity: "Entität",
-  editor_name: "Name",
-  editor_color: "Farbe",
-  editor_view_start_time: "Anfangszeit",
-  editor_view_end_time: "Endzeit",
-  editor_view_slot_minutes: "Zeitslot-Minuten",
-  editor_locale: "Sprache",
-  editor_time_format: "Zeitformat",
-  editor_show_now_indicator: "Jetzt-Indikator anzeigen",
-  editor_height_vh: "Höhe (vh)",
-  editor_px_per_min: "Pixel pro Minute",
-  editor_remember_offset: "Versatz merken",
-  editor_data_refresh_minutes: "Datenaktualisierung (Minuten)",
-  editor_weather_entity: "Wetterentität",
-  editor_remove: "Entfernen",
 } as const;
