+++ src/locales/en.ts
@@ -10,21 +10,4 @@ export default {
   aria_next_week: "Next week",
   aria_today: "Go to current week",
   failed_load_prefix: "Failed to load:",
-  editor_calendars: "Calendars",
-  editor_add_calendar: "Add calendar",
-  editor_entity: "Entity",
-  editor_name: "Name",
-  editor_color: "Color",
-  editor_view_start_time: "View start time",
-  editor_view_end_time: "View end time",
-  editor_view_slot_minutes: "View slot minutes",
-  editor_locale: "Locale",
-  editor_time_format: "Time format",
-  editor_show_now_indicator: "Show now indicator",
-  editor_height_vh: "Height (vh)",
-  editor_px_per_min: "Pixels per minute",
-  editor_remember_offset: "Remember offset",
-  editor_data_refresh_minutes: "Data refresh minutes",
-  editor_weather_entity: "Weather entity",
-  editor_remove: "Remove",
 } as const;
